refactor: move EnglishResult to search.py
